# docs/openapi.yaml
openapi: 3.1.0
info: { title: Aiden Support API, version: 1.0.0 }
servers: [{ url: http://localhost:3000 }]
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
paths:
...
  /api/organizations:
    get: { summary: List memberships, security: [{ bearerAuth: [] }], responses: { '200': { description: Memberships } } }
    post: { summary: Create organization, security: [{ bearerAuth: [] }], responses: { '200': { description: Organization } } }
...